refactor: add consistent pattern
