+++ tests/support/data/licence.js
@@ -1,6 +1,6 @@
 import { randomUUID } from 'crypto'
 
-export default function (licenceRef, companies, primaryUser = null) {
+export default function (licenceRef, companiesData, primaryUserData = null) {
   const licenceDocumentHeaderId = randomUUID()
   const licenceDocumentId = randomUUID()
   const pointId = randomUUID()
@@ -8,12 +8,12 @@ export default function (licenceRef, companies, primaryUser = null) {
   const licenceVersionId = randomUUID()
   const licenceVersionPurposeId = randomUUID()
 
-  const company = companies.companies[0]
-  const address = companies.addresses[0]
+  const company = companiesData.companies[0]
+  const address = companiesData.addresses[0]
 
   // When there is a primary user, we need to link them to the 'licenceDocumentHeaders'; this is the only way we can
   // link a registered licence to a licence holder.
-  const companyEntityId = primaryUser ? primaryUser.licenceEntityRoles[0].companyEntityId : null
+  const companyEntityId = primaryUserData ? primaryUserData.licenceEntityRoles[0].companyEntityId : null
 
   return {
     permitLicences: [
